refactor: 훅 분리 및 파일 분리
