+++ src/pages/GraduationCalculatorPage/components/CourseTable/EllipsisTooltip/EllipsisTooltip.module.scss
